Refactor IdentifierType
