+++ app/schemas/contributor.py
@@ -6,12 +6,14 @@ from uuid import UUID
 from entitysdk.types import AgentType
 from pydantic import BaseModel, Field
 
+from app.types import IdentifierType
+
 
 class PersonPreview(BaseModel):
     """Preview of a person resolved from an ORCID."""
 
     identifier: str
-    identifier_type: Literal["orcid"] = "orcid"
+    identifier_type: Literal[IdentifierType.orcid] = IdentifierType.orcid
     agent_type: Literal[AgentType.person] = AgentType.person
     name: str
     given_name: str | None = None
@@ -25,7 +27,7 @@ class OrganizationPreview(BaseModel):
     """Preview of an organization resolved from a ROR ID."""
 
     identifier: str
-    identifier_type: Literal["ror"] = "ror"
+    identifier_type: Literal[IdentifierType.ror] = IdentifierType.ror
     agent_type: Literal[AgentType.organization] = AgentType.organization
     name: str
     alternative_name: str | None = None
@@ -38,3 +40,22 @@ ContributorPreview = Annotated[
     PersonPreview | OrganizationPreview,
     Field(discriminator="agent_type"),
 ]
+
+
+class OrcidMetadata(BaseModel):
+    """Metadata for a person fetched from the ORCID API."""
+
+    orcid: str
+    given_name: str | None = None
+    family_name: str | None = None
+    pref_label: str
+
+
+class RorMetadata(BaseModel):
+    """Metadata for an organization fetched from the ROR API."""
+
+    ror_id: str
+    name: str
+    alternative_names: list[str] = []
+    types: list[str] = []
+    country: str | None = None
